+++ Example/Source/Core/UIViewControllers/RootUINavigationController.swift
@@ -17,4 +17,8 @@ class RootUINavigationController : UINavigationController {
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+    
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return topViewController?.preferredStatusBarStyle ?? .default
+    }
 }