+++ Example/Source/Protocols/ContentActionable.swift
@@ -9,9 +9,11 @@
 protocol ContentActionable {
     typealias Action = () -> Void
     
+    var isShowAdvancedActions: Bool { get set }
     var removeDidTapAction: Action? { get set }
     var insertDidTapAction: Action? { get set }
     var appendDidTapAction: Action? { get set }
     var menuDidTapAction: Action? { get set }
     var changePositionAction: ((Int) -> ())? { get set }
+    var changeTitleModeAction: ((Int) -> ())? { get set }
 }